create merge_intervals util func
